Tried to separate processes monitor and hub but not working
