+++ HubHelpers.h
@@ -1,0 +1,11 @@
+#ifndef HUBHELPERS_H
+#define HUBHELPERS_H
+
+void executeOperation(const char *operation);
+
+void initTreasureHubProcess(void);
+
+void startMonitor(void);
+
+void exitHub(void);
+#endif //HUBHELPERS_H
