+++ MonitorHelpers.h
@@ -1,0 +1,6 @@
+#ifndef MONITORHELPERS_H
+#define MONITORHELPERS_H
+
+void initMonitorProcess(void);
+
+#endif //MONITORHELPERS_H
