+++ operationHelpers.c
@@ -2,12 +2,9 @@
 #include <signal.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <string.h>
-#include <unistd.h>
-#include "treasureMonitor.h"
-#include "FileLib.h"
 #include "DirectoryLib.h"
-#include "Treasure.h"
+#include "FileLib.h"
+#include <string.h>
 
 static char (*splitArguments(char *operation))[1024] {//Split the arguments to mimic the argv format
     static char arguments[4][1024]; //The operations that we should be able to parse have at most 3 parameters. First one will be the argCount so that we know without going over it again
@@ -20,7 +17,7 @@ static char (*splitArguments(char *operation))[1024] {//Split the arguments to m
     size_t cnt = 1; //postiion of the arguments parsed
     while (param != NULL) {
         strcpy(arguments[cnt++], param);
-        if (cnt > 3) {
+        if (cnt > 4) {
             fprintf(stderr, "Too many arguments\n");
             return NULL;
         }
@@ -28,29 +25,15 @@ static char (*splitArguments(char *operation))[1024] {//Split the arguments to m
     }
 
     sprintf(arguments[0], "%ld", cnt);
-
+    //printf("%s\n%s\n%s\n%s\n", arguments[0], arguments[1], arguments[2], arguments[3]);
     return arguments;
 }
 
-static void parseOperation(const char * operation, const char (*parameters)[1024], const size_t parametersCount) {
-    if (strcmp(operation, "stop_monitor") == 0) {
-        stopMonitor();
-        return;
+void notifyProcess(const int processPid, const int signal) {
+    if (kill(processPid, signal) < 0) {//SIGUSR1 means that an operaion is recieved
+        perror("Impossible to send a signal to monitor");
+        exit(3);
     }
-    else if (strcmp(operation, "list_hunts") == 0) {
-        listHunts();
-    }
-    else if (strcmp(operation, "list_treasures") == 0) {
-        listTreasure(parameters[0]);
-    }
-    else if (strcmp(operation, "view_treasure") == 0) {
-        viewTreasure(parameters[0], parameters[1]);
-    }
-    else {
-         printf("\033[31mUnknown operation: %s\nUsage : \nstart_monitor\nlist_hunts\nlist_treasure <hunt_id>\nview_treasure <hunt_id> <treasure_id>\nstop_monitor\nexit\n\033[0m", operation);
-        fflush(stdout);
-    }
-    notifyProcess(getppid(), SIGUSR2); //Notify the parrent that we finished parsing this, so we can continue
 }
 
 void writeOperationInfoInFile(const char *operation) {
@@ -67,14 +50,7 @@ void writeOperationInfoInFile(const char *operation) {
     closeFile(fd);
 }
 
-void notifyProcess(const int processPid, const int signal) {
-    if (kill(processPid, signal) < 0) {//SIGUSR1 means that an operaion is recieved
-        perror("Impossible to send a signal to monitor");
-        exit(3);
-    }
-}
-
-void readOperationInfoFromFile(int sig) {
+char (*readOperationInfoFromFile(void))[1024] {
     //If the signal SIGUSR1 is recieved we are sure that the monitor process is running
     int fd = openFile("insides/operations.in", "r");
     char operationBuffer[1024] = {0};
@@ -83,10 +59,6 @@ void readOperationInfoFromFile(int sig) {
     closeFile(fd);
 
     //After reading the operation we need to parse it so that we make it the same way command line arguments are
-    char (*operationParameters)[1024] = splitArguments(operationBuffer);
 
-    if (operationParameters != NULL) {
-        parseOperation(operationParameters[1], operationParameters + 2, strtol(operationParameters[0], NULL, 10));
-        //Parse the operation
-    }
+    return splitArguments(operationBuffer);
 }
